refactoring : 리팩토링 로직 분리
